+++ src/command_modules/azure-cli-appservice/HISTORY.rst
@@ -4,6 +4,7 @@ Release History
 ===============
 0.2.4
 +++++
+* support for webjobs(continuous and triggered) operations management
 * appservice plan, webapp & function app updated to use latest python websites SDK version
 * az webapp config set supports --fts-state property. Also added support fot az functionapp config set & show
 * webapp: add support for bring your own storage